services/api/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
asyncpg==0.30.*
alembic==1.14.*
pgvector==0.3.*
pydantic==2.10.*
pydantic-settings==2.7.*
python-multipart==0.0.*
httpx==0.28.*
structlog==24.*
pypdf==5.*
pyyaml==6.*
pyjwt==2.*
